Move components -> elements
