+++ packages/typography/src/elements/Ellipsis.tsx
