+++ packages/typography/src/elements/LG.tsx
