+++ packages/typography/src/utils/useOrderedListContext.spec.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import { render } from 'garden-test-utils';
-import OrderedList from '../components/lists/OrderedList';
+import OrderedList from '../elements/lists/OrderedList';
 import useOrderedListContext from './useOrderedListContext';
 
 describe('useOrderedListContext', () => {
